+++ env_canada/constants.py
@@ -1,1 +1,1 @@
-USER_AGENT = "env_canada/0.14.1"
+USER_AGENT = "env_canada/0.15.0"